Move files to internal model directory
