+++ internal/model/commit-message_test.go
@@ -1,4 +1,4 @@
-package src
+package model
 
 import (
 	"github.com/stretchr/testify/assert"
